# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
  system: 'system',
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface RoleplayMessageRequest {
  scenarioId: string;
  messages: ChatMessage[];
  /** Optional custom AI system prompt for org-specific scenarios. When provided, overrides the built-in prompt for the given scenarioId.
   */
  customSystemPrompt?: string;
}

export interface RoleplayMessageResponse {
  response: string;
}

export interface RoleplayFeedbackRequest {
  scenarioId: string;
  messages: ChatMessage[];
  /** Optional custom scenario context for more accurate feedback. */
  customSystemPrompt?: string;
}

export interface TranscribeRequest {
  /** Base64-encoded audio (webm/ogg/mp4/wav/m4a). */
  audioBase64: string;
  /** MIME type of the audio, e.g. "audio/webm". */
  mimeType: string;
}

export interface TranscribeResponse {
  text: string;
}

export interface SpeechRequest {
  text: string;
  /** Voice id (alloy, echo, fable, onyx, nova, shimmer). Defaults to "nova". */
  voice?: string;
}

export interface SpeechResponse {
  /** Base64-encoded MP3 audio. */
  audioBase64: string;
  mimeType: string;
}

export interface RoleplayFeedbackResponse {
  score: number;
  summary: string;
  strengths: string[];
  improvements: string[];
}