# lib/api-zod/src/generated/types/roleplayFeedbackRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from './chatMessage';

export interface RoleplayFeedbackRequest {
  scenarioId: string;
  messages: ChatMessage[];
  /** Optional custom scenario context for more accurate feedback. */
  customSystemPrompt?: string;
}