# lib/api-zod/src/generated/types/roleplayMessageRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from './chatMessage';

export interface RoleplayMessageRequest {
  scenarioId: string;
  messages: ChatMessage[];
  /** Optional custom AI system prompt for org-specific scenarios. When provided, overrides the built-in prompt for the given scenarioId.
   */
  customSystemPrompt?: string;
}