# lib/api-zod/src/generated/types/speechRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SpeechRequest {
  text: string;
  /** Voice id (alloy, echo, fable, onyx, nova, shimmer). Defaults to "nova". */
  voice?: string;
}