# lib/api-zod/src/generated/types/speechResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SpeechResponse {
  /** Base64-encoded MP3 audio. */
  audioBase64: string;
  mimeType: string;
}