# lib/api-zod/src/generated/types/transcribeRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TranscribeRequest {
  /** Base64-encoded audio (webm/ogg/mp4/wav/m4a). */
  audioBase64: string;
  /** MIME type of the audio, e.g. "audio/webm". */
  mimeType: string;
}